Convert SelectionBox to typescript
Need to look into how it works to figure out a certain case when trying to select an extra group of resources.

Nothing special needed in the translaton. Was matter of changing name
